Adjust LeftPane width in landscape mode for Login and Registration
The width of the `LeftPane` in landscape mode for both the Login and Registration screens has been changed from 50% (`fillMaxWidth(0.5f)`) to 40% (`fillMaxWidth(0.4f)`) of the screen width.

Fix paddings in Login and Registration Screens

This commit adjusts the padding in the `LoginPane` and `RegistrationPane` composables.

In portrait mode, the start and end paddings were removed from the background image to allow it to extend to the screen edges.

In landscape mode, the start and end paddings for the `ForegroundLoginPane` and `ForegroundRegistrationPane` were updated to include display cutouts, ensuring content is not obscured.

Additionally, a fixed height of 32.dp was added to the `Spacer` with `imePadding` in `ForegroundRegistrationPane` to provide consistent spacing when the keyboard is visible.