Enable and implement periodic work for note synchronization
This commit enables the periodic synchronization of notes by uncommenting the relevant code in `AppStateMachine.kt`. The `cancelPreviousAndTriggerNewWork()` method is now called to schedule or reschedule the synchronization work when the user logs in or when the sync duration is changed in the settings.

Additionally, the following changes were made:
- In `DateTime.kt`, the fallback return value for date/time formatting functions in case of an exception has been changed from an empty string to "Unknown".
- A new composable `SyncDropDown.kt` has been created by extracting the dropdown menu logic from `SettingsPane.kt`. This improves code organization and reusability. `SettingsPane.kt` now utilizes this new `SyncDropDown` composable.